+++ operator/CHANGELOG.md
@@ -12,9 +12,11 @@ We structure this changelog in accordance with [Keep a Changelog](https://keepac
 
 ### Added
 
+- Add target PVC claim template to the ais-target configmap. This can be used by AIS ETL to mount specific data PVCs for FQN. 
+
 ### Changed
 
-- Update RBAC rule in AIS service account to remove access to secrets amd configmaps and add GET access to PVCs
+- Update RBAC rule in AIS service account to remove access to secrets and configmaps and add GET access to PVCs
 
 ## v2.3.0